+++ CHANGELOG.md
@@ -65,6 +65,7 @@
     * Call action `admin-pre-header` only on inner administration panel pages
     * Delete autosave copy when deleting page
     * Query parameters are preserved on canonical redirect
+    * Signature of function `get_navigation()` changed. Added parameter to return code of the menu as string
 
 1. [](#bugfix)
     * `index` page now processed with **Prerry URLs** as other pages. To preserve compatibility user can set custom permalink structure on `index` page to `/`